Refactor: change directory
