+++ app/register/steps/step4.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/navigation';
-import { FinishIcon } from '@/icons';
+import { FinishIcon } from '@/lib/icons';
 import { HStack } from '@/components/basic/stack';
 import Button from '@/components/basic/button';
