+++ components/basic/checkbok.tsx
@@ -1,4 +1,4 @@
-import { CheckedIcon, UncheckedIcon } from '@/icons';
+import { CheckedIcon, UncheckedIcon } from '@/lib/icons';
 
 interface CheckboxProps {
   checked: boolean;
