+++ lib/icons/warning.tsx
